+++ spinta/cli/sync.py
@@ -6,8 +6,8 @@ from typer import Context as TyperContext, Argument, Option
 
 from spinta.cli.helpers.manifest import convert_str_to_manifest_path
 from spinta.cli.helpers.sync import ContentType
+from spinta.cli.helpers.sync.controllers.data_service import get_data_service_id
 from spinta.cli.helpers.sync.controllers.dataset import get_resource, create_resource
-from spinta.cli.helpers.sync.controllers.distribution import create_distribution
 from spinta.cli.helpers.sync.controllers.dsa import update_dsa, create_dsa
 from spinta.cli.helpers.sync.helpers import (
     get_base_path_and_headers,
@@ -17,6 +17,8 @@ from spinta.cli.helpers.sync.helpers import (
     extract_identifier_from_response,
     get_configuration_credentials,
     get_data_service_name_prefix,
+    get_agent_name,
+    validate_credentials,
 )
 
 logger = logging.getLogger(__name__)
@@ -67,11 +69,12 @@ def sync(
     dataset_data = prepare_synchronization_manifests(context, manifest)
 
     credentials = get_configuration_credentials(context)
+    validate_credentials(credentials)
     base_path, headers = get_base_path_and_headers(credentials)
+    agent_name = get_agent_name(credentials)
     prefix = get_data_service_name_prefix(credentials)
 
-    response_create_data_service = create_resource(base_path, headers, f"{prefix}", resource_type="service")
-    data_service_id = extract_identifier_from_response(response_create_data_service, "detail")
+    data_service_id = get_data_service_id(base_path, headers, agent_name)
     for dataset in dataset_data:
         dataset_name = f"{prefix}/{dataset['name']}"
         response_get_dataset = get_resource(base_path, headers, dataset_name)
@@ -81,10 +84,5 @@ def sync(
         elif response_get_dataset.status_code == HTTPStatus.NOT_FOUND:
             response_create_dataset = create_resource(base_path, headers, dataset_name, data_service_id)
             dataset_id = extract_identifier_from_response(response_create_dataset, "detail")
-            for distribution in dataset["resources"]:
-                file_bytes = render_content_from_manifest(context, distribution["manifest"], ContentType.BYTES)
-                create_distribution(
-                    base_path, headers, f"{dataset_name}/{distribution['name']}", file_bytes, dataset_id
-                )
             content = render_content_from_manifest(context, dataset["dataset_manifest"], ContentType.CSV)
             create_dsa(base_path, headers, dataset_id, content)