+++ tests/test_sync.py
@@ -48,6 +48,7 @@ def patched_credentials():
         client="client-id",
         secret="secret",
         server="http://example.com",
+        resource_server="http://example.com",
         scopes="scope1 scope2",
         organization="vssa",
         organization_type="gov",
@@ -125,8 +126,10 @@ def test_success_existing_dataset(
     )
     mock_dataset_get = requests_mock.get(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        status_code=HTTPStatus.OK,
-        json={"_data": [{"_id": 2}]},
+        [
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+            {"status_code": HTTPStatus.OK, "json": {"_data": [{"_id": 2}]}},
+        ],
     )
     mock_dataset_put = requests_mock.put(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/2/dsa/",
@@ -134,6 +137,7 @@ def test_success_existing_dataset(
         json={},
     )
     dataset_name = f"{dataset_prefix}/example"
+    agent_name = patched_credentials.client
 
     # Act
     with pytest.raises(NotImplementedFeature) as exception:
@@ -164,8 +168,8 @@ def test_success_existing_dataset(
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "subclass": [ResourceType.DATA_SERVICE.value],
                 "service": ["True"],
             },
@@ -174,10 +178,16 @@ def test_success_existing_dataset(
     assert get_request_context(mock_dataset_get) == [
         {
             "method": "GET",
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
+            "data": {},
+        },
+        {
+            "method": "GET",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(dataset_name, safe='')}",
             "params": {"name": [dataset_name]},
             "data": {},
-        }
+        },
     ]
     assert get_request_context(mock_dataset_put) == [
         {
@@ -207,8 +217,10 @@ def test_success_new_dataset(
     )
     mock_dataset_get = requests_mock.get(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        status_code=HTTPStatus.NOT_FOUND,
-        json={},
+        [
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+        ],
     )
     mock_dataset_post = requests_mock.post(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
@@ -217,11 +229,6 @@ def test_success_new_dataset(
             {"status_code": HTTPStatus.CREATED, "json": {"_id": 2}},  # Creates Dataset No. 1.
             {"status_code": HTTPStatus.CREATED, "json": {"_id": 3}},  # Creates Dataset No. 2.
         ],
-    )
-    mock_distribution_post = requests_mock.post(
-        f"{patched_credentials.server}/{base_uapi_url}/Distribution/",
-        status_code=HTTPStatus.CREATED,
-        json={"_id": 1},
     )
     mock_dsa_post = requests_mock.post(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/2/dsa/",
@@ -233,6 +240,8 @@ def test_success_new_dataset(
         status_code=HTTPStatus.NO_CONTENT,
         json={},
     )
+
+    agent_name = patched_credentials.client
     dataset_name_example = f"{dataset_prefix}/example"
     dataset_name_sqlite = f"{dataset_prefix}/db_sqlite"
 
@@ -254,6 +263,12 @@ def test_success_new_dataset(
     assert get_request_context(mock_dataset_get) == [
         {
             "method": "GET",
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
+            "data": {},
+        },
+        {
+            "method": "GET",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(dataset_name_example, safe='')}",
             "params": {"name": [f"{dataset_prefix}/example"]},
             "data": {},
@@ -271,8 +286,8 @@ def test_success_new_dataset(
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "service": ["True"],
                 "subclass": [ResourceType.DATA_SERVICE],
             },
@@ -298,20 +313,6 @@ def test_success_new_dataset(
                 "parent_id": ["1"],
                 "subclass": [ResourceType.DATASET],
             },
-        },
-    ]
-    assert get_request_context(mock_distribution_post) == [
-        {
-            "method": "POST",
-            "url": f"{patched_credentials.server}/{base_uapi_url}/Distribution/",
-            "params": {},
-            "data": ANY,  # DSA content + SQLite content.
-        },
-        {
-            "method": "POST",
-            "url": f"{patched_credentials.server}/{base_uapi_url}/Distribution/",
-            "params": {},
-            "data": ANY,  # DSA content + SQLite content.
         },
     ]
     assert get_request_context(mock_dsa_post, with_text=True) == [
@@ -361,14 +362,13 @@ def test_failure_configuration_invalid(
     dataset_prefix: str,
 ):
     # Arrange
-    # No `organization` or `organization_type`
     credentials = RemoteClientCredentials(
         section="default",
         remote="origin",
-        client="client-id",
         secret="secret",
-        server="http://example.com",
         scopes="scope1 scope2",
+        client="client",
+        server="https://example.com",
     )
     requests_mock.post(
         f"{credentials.server}/auth/token",
@@ -382,7 +382,7 @@ def test_failure_configuration_invalid(
 
         # Assert
         assert exception.value.status_code == 400
-        assert exception.value.context == {"required_credentials": str(["organization_type", "organization"])}
+        assert exception.value.context == {"missing_credentials": "resource_server, organization_type, organization"}
 
 
 def test_failure_get_access_token_api_call(
@@ -437,11 +437,18 @@ def test_failure_post_data_service_returns_unexpected_status_code(
         status_code=HTTPStatus.OK,
         json={"access_token": "test-token"},
     )
+    mock_dataset_get = requests_mock.get(
+        f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
+        status_code=HTTPStatus.NOT_FOUND,
+        json={},
+    )
     mock_data_service_post = requests_mock.post(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
         status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
         json={},
     )
+
+    agent_name = patched_credentials.client
 
     # Act
     with pytest.raises(UnexpectedAPIResponse) as exception:
@@ -467,14 +474,22 @@ def test_failure_post_data_service_returns_unexpected_status_code(
             },
         }
     ]
+    assert get_request_context(mock_dataset_get) == [
+        {
+            "method": "GET",
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
+            "data": {},
+        },
+    ]
     assert get_request_context(mock_data_service_post) == [
         {
             "method": "POST",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "subclass": [ResourceType.DATA_SERVICE.value],
                 "service": ["True"],
             },
@@ -498,11 +513,18 @@ def test_failure_post_data_service_returns_invalid_data(
         status_code=HTTPStatus.OK,
         json={"access_token": "test-token"},
     )
+    mock_data_service_get = requests_mock.get(
+        f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
+        status_code=HTTPStatus.NOT_FOUND,
+        json={},
+    )
     mock_data_service_post = requests_mock.post(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
         status_code=HTTPStatus.CREATED,
         json={},
     )
+
+    agent_name = patched_credentials.client
 
     # Act
     with pytest.raises(UnexpectedAPIResponseData) as exception:
@@ -526,14 +548,22 @@ def test_failure_post_data_service_returns_invalid_data(
             },
         }
     ]
+    assert get_request_context(mock_data_service_get) == [
+        {
+            "method": "GET",
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
+            "data": {},
+        },
+    ]
     assert get_request_context(mock_data_service_post) == [
         {
             "method": "POST",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "subclass": [ResourceType.DATA_SERVICE.value],
                 "service": ["True"],
             },
@@ -564,16 +594,23 @@ def test_failure_get_dataset_returns_unexpected_status_code(
     )
     mock_dataset_get = requests_mock.get(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
-        json={
-            "code": "dataset_not_found",
-            "type": "DatasetNotFound",
-            "template": "The requested Dataset could not be found.",
-            "message": "No dataset matched the provided query.",
-            "status_code": HTTPStatus.INTERNAL_SERVER_ERROR.value,
-        },
+        [
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+            {
+                "status_code": HTTPStatus.INTERNAL_SERVER_ERROR,
+                "json": {
+                    "code": "dataset_not_found",
+                    "type": "DatasetNotFound",
+                    "template": "The requested Dataset could not be found.",
+                    "message": "No dataset matched the provided query.",
+                    "status_code": HTTPStatus.INTERNAL_SERVER_ERROR.value,
+                },
+            },
+        ],
     )
+
     dataset_name_example = f"{dataset_prefix}/example"
+    agent_name = patched_credentials.client
 
     # Act
     with pytest.raises(UnexpectedAPIResponse) as exception:
@@ -612,8 +649,8 @@ def test_failure_get_dataset_returns_unexpected_status_code(
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "subclass": [ResourceType.DATA_SERVICE.value],
                 "service": ["True"],
             },
@@ -622,10 +659,16 @@ def test_failure_get_dataset_returns_unexpected_status_code(
     assert get_request_context(mock_dataset_get) == [
         {
             "method": "GET",
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
+            "data": {},
+        },
+        {
+            "method": "GET",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(dataset_name_example, safe='')}",
             "params": {"name": [dataset_name_example]},
             "data": {},
-        }
+        },
     ]
 
 
@@ -652,9 +695,13 @@ def test_failure_get_dataset_returns_invalid_data(
     )
     mock_dataset_get = requests_mock.get(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        status_code=HTTPStatus.OK,
-        json={},
+        [
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+            {"status_code": HTTPStatus.OK, "json": {}},
+        ],
     )
+
+    agent_name = patched_credentials.client
     dataset_name_example = f"{dataset_prefix}/example"
 
     # Act
@@ -685,8 +732,8 @@ def test_failure_get_dataset_returns_invalid_data(
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "subclass": [ResourceType.DATA_SERVICE.value],
                 "service": ["True"],
             },
@@ -695,10 +742,16 @@ def test_failure_get_dataset_returns_invalid_data(
     assert get_request_context(mock_dataset_get) == [
         {
             "method": "GET",
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
+            "data": {},
+        },
+        {
+            "method": "GET",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(dataset_name_example, safe='')}",
             "params": {"name": [dataset_name_example]},
             "data": {},
-        }
+        },
     ]
 
 
@@ -730,13 +783,17 @@ def test_failure_put_dataset_returns_invalid_data(
     )
     mock_dataset_get = requests_mock.get(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        status_code=HTTPStatus.OK,
-        json={"_data": [{"_id": 2}]},
+        [
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+            {"status_code": HTTPStatus.OK, "json": {"_data": [{"_id": 2}]}},
+        ],
     )
     mock_dsa_put = requests_mock.put(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/2/dsa/",
         status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
     )
+
+    agent_name = patched_credentials.client
     dataset_name_example = f"{dataset_prefix}/example"
 
     # Act
@@ -768,8 +825,8 @@ def test_failure_put_dataset_returns_invalid_data(
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "subclass": [ResourceType.DATA_SERVICE.value],
                 "service": ["True"],
             },
@@ -778,10 +835,16 @@ def test_failure_put_dataset_returns_invalid_data(
     assert get_request_context(mock_dataset_get) == [
         {
             "method": "GET",
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
+            "data": {},
+        },
+        {
+            "method": "GET",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(dataset_name_example, safe='')}",
             "params": {"name": [dataset_name_example]},
             "data": {},
-        }
+        },
     ]
     assert get_request_context(mock_dsa_put) == [
         {
@@ -818,9 +881,13 @@ def test_failure_post_dataset_returns_unexpected_status_code(
     )
     mock_dataset_get = requests_mock.get(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        status_code=HTTPStatus.NOT_FOUND,
-        json={},
+        [
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+        ],
     )
+
+    agent_name = patched_credentials.client
     dataset_name_example = f"{dataset_prefix}/example"
 
     # Act
@@ -853,8 +920,8 @@ def test_failure_post_dataset_returns_unexpected_status_code(
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "subclass": [ResourceType.DATA_SERVICE.value],
                 "service": ["True"],
             },
@@ -874,10 +941,16 @@ def test_failure_post_dataset_returns_unexpected_status_code(
     assert get_request_context(mock_dataset_get) == [
         {
             "method": "GET",
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
+            "data": {},
+        },
+        {
+            "method": "GET",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(dataset_name_example, safe='')}",
             "params": {"name": [dataset_name_example]},
             "data": {},
-        }
+        },
     ]
 
 
@@ -899,8 +972,10 @@ def test_failure_post_dataset_returns_invalid_data(
     )
     mock_dataset_get = requests_mock.get(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        status_code=HTTPStatus.NOT_FOUND,
-        json={},
+        [
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+        ],
     )
     mock_dataset_post = requests_mock.post(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
@@ -909,6 +984,8 @@ def test_failure_post_dataset_returns_invalid_data(
             {"status_code": HTTPStatus.CREATED, "json": {}},
         ],
     )
+
+    agent_name = patched_credentials.client
     dataset_name_example = f"{dataset_prefix}/example"
 
     # Act
@@ -939,8 +1016,8 @@ def test_failure_post_dataset_returns_invalid_data(
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "subclass": [ResourceType.DATA_SERVICE.value],
                 "service": ["True"],
             },
@@ -960,110 +1037,15 @@ def test_failure_post_dataset_returns_invalid_data(
     assert get_request_context(mock_dataset_get) == [
         {
             "method": "GET",
-            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(dataset_name_example, safe='')}",
-            "params": {"name": [dataset_name_example]},
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
             "data": {},
-        }
-    ]
-
-
-def test_failure_post_distribution_returns_unexpected_status_code(
-    rc: RawConfig,
-    cli: SpintaCliRunner,
-    manifest_path: PosixPath,
-    requests_mock: MagicMock,
-    patched_credentials: RemoteClientCredentials,
-    base_uapi_url: str,
-    sqlite_instance: Sqlite,
-    dataset_prefix: str,
-):
-    # Arrange
-    mock_auth_token_post = requests_mock.post(
-        f"{patched_credentials.server}/auth/token",
-        status_code=HTTPStatus.OK,
-        json={"access_token": "test-token"},
-    )
-    mock_dataset_get = requests_mock.get(
-        f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        status_code=HTTPStatus.NOT_FOUND,
-        json={},
-    )
-    mock_dataset_post = requests_mock.post(
-        f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        [
-            {"status_code": HTTPStatus.CREATED, "json": {"_id": 1}},
-            {"status_code": HTTPStatus.CREATED, "json": {"_id": 2}},
-        ],
-    )
-    mock_distribution_post = requests_mock.post(
-        f"{patched_credentials.server}/{base_uapi_url}/Distribution/",
-        status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
-        json={},
-    )
-    dataset_name_example = f"{dataset_prefix}/example"
-
-    # Act
-    with pytest.raises(UnexpectedAPIResponse) as exception:
-        cli.invoke(rc, args=["sync", manifest_path, "-r", "sql", sqlite_instance.dsn], catch_exceptions=False)
-
-    # Assert
-    assert exception.value.status_code == HTTPStatus.INTERNAL_SERVER_ERROR.value
-    assert exception.value.context == {
-        "operation": "Create distribution",
-        "expected_status_code": str({HTTPStatus.CREATED.value}),
-        "response_status_code": HTTPStatus.INTERNAL_SERVER_ERROR.value,
-        "response_data": str({}),
-    }
-
-    assert get_request_context(mock_auth_token_post) == [
-        {
-            "method": "POST",
-            "url": f"{patched_credentials.server}/auth/token",
-            "params": {},
-            "data": {
-                "grant_type": ["client_credentials"],
-                "scope": [patched_credentials.scopes],
-            },
-        }
-    ]
-    assert get_request_context(mock_dataset_get) == [
+        },
         {
             "method": "GET",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(dataset_name_example, safe='')}",
             "params": {"name": [dataset_name_example]},
             "data": {},
-        }
-    ]
-    assert get_request_context(mock_dataset_post) == [
-        {
-            "method": "POST",
-            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-            "params": {},
-            "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
-                "subclass": [ResourceType.DATA_SERVICE.value],
-                "service": ["True"],
-            },
-        },
-        {
-            "method": "POST",
-            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-            "params": {},
-            "data": {
-                "name": [dataset_name_example],
-                "title": [dataset_name_example],
-                "parent_id": ["1"],
-                "subclass": [ResourceType.DATASET],
-            },
-        },
-    ]
-    assert get_request_context(mock_distribution_post) == [
-        {
-            "method": "POST",
-            "url": f"{patched_credentials.server}/{base_uapi_url}/Distribution/",
-            "params": {},
-            "data": ANY,  # DSA content + SQLite content.
         },
     ]
 
@@ -1086,8 +1068,10 @@ def test_failure_post_dsa_returns_unexpected_status_code(
     )
     mock_dataset_get = requests_mock.get(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
-        status_code=HTTPStatus.NOT_FOUND,
-        json={},
+        [
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+            {"status_code": HTTPStatus.NOT_FOUND, "json": {}},
+        ],
     )
     mock_dataset_post = requests_mock.post(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
@@ -1096,16 +1080,13 @@ def test_failure_post_dsa_returns_unexpected_status_code(
             {"status_code": HTTPStatus.CREATED, "json": {"_id": 2}},
         ],
     )
-    mock_distribution_post = requests_mock.post(
-        f"{patched_credentials.server}/{base_uapi_url}/Distribution/",
-        status_code=HTTPStatus.CREATED,
-        json={"_id": 1},
-    )
     mock_dsa_post = requests_mock.post(
         f"{patched_credentials.server}/{base_uapi_url}/Dataset/2/dsa/",
         status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
         json={},
     )
+
+    agent_name = patched_credentials.client
     dataset_name_example = f"{dataset_prefix}/example"
 
     # Act
@@ -1135,10 +1116,16 @@ def test_failure_post_dsa_returns_unexpected_status_code(
     assert get_request_context(mock_dataset_get) == [
         {
             "method": "GET",
+            "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(agent_name, safe='')}",
+            "params": {"name": [agent_name]},
+            "data": {},
+        },
+        {
+            "method": "GET",
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/?name={quote(dataset_name_example, safe='')}",
             "params": {"name": [dataset_name_example]},
             "data": {},
-        }
+        },
     ]
     assert get_request_context(mock_dataset_post) == [
         {
@@ -1146,8 +1133,8 @@ def test_failure_post_dsa_returns_unexpected_status_code(
             "url": f"{patched_credentials.server}/{base_uapi_url}/Dataset/",
             "params": {},
             "data": {
-                "name": [dataset_prefix],
-                "title": [dataset_prefix],
+                "name": [patched_credentials.client],
+                "title": [patched_credentials.client],
                 "subclass": [ResourceType.DATA_SERVICE.value],
                 "service": ["True"],
             },
@@ -1162,14 +1149,6 @@ def test_failure_post_dsa_returns_unexpected_status_code(
                 "parent_id": ["1"],
                 "subclass": [ResourceType.DATASET],
             },
-        },
-    ]
-    assert get_request_context(mock_distribution_post) == [
-        {
-            "method": "POST",
-            "url": f"{patched_credentials.server}/{base_uapi_url}/Distribution/",
-            "params": {},
-            "data": ANY,  # DSA content + SQLite content.
         },
     ]
     assert get_request_context(mock_dsa_post) == [